Add background audio mute preference
